+++ tests/custom_command_formatting/keywords_from_variables.cmake
@@ -1,4 +1,4 @@
-function(SEVEN_SAMURAI some standalone arguments)
+function(Seven_Samurai some standalone arguments)
     set(options KAMBEI KATSUSHIRO)
     set(oneValueArgs GOROBEI HEIHACHI KYUZO)
     set(multiValueArgs SHICHIROJI KIKUCHIYO)