+++ tests/custom_command_formatting/keywords_in_outer_scope.cmake
@@ -2,7 +2,7 @@ set(multiValueArgs SHICHIROJI KIKUCHIYO)
 set(oneValueArgs GOROBEI HEIHACHI KYUZO)
 set(options KAMBEI KATSUSHIRO)
 
-function(SEVEN_SAMURAI some standalone arguments)
+function(Seven_Samurai some standalone arguments)
     cmake_parse_arguments(
         THIS_FUNCTION_PREFIX
         "${options}"