+++ tests/formatter/downcase_unknown_command_invocation_identifier.in.cmake
@@ -1,0 +1,8 @@
+foobar_set(FOO BAR)
+
+Foobar_Set(FOO BAR)
+
+FOOBAR_SET(FOO BAR)
+
+fOOBAR_sET(FOO BAR)
+