+++ tests/formatter/downcase_unknown_command_invocation_identifier.out.cmake
@@ -1,0 +1,7 @@
+foobar_set(FOO BAR)
+
+foobar_set(FOO BAR)
+
+foobar_set(FOO BAR)
+
+foobar_set(FOO BAR)