+++ tests/formatter/issue_0021_builtin_commands_with_non_lower_case_canonical_names.in.cmake
@@ -1,0 +1,3 @@
+fetchcontent_declare( gersemi GIT_REPOSITORY https://github.com/BlankSpruce/gersemi.git GIT_TAG abcdefabcdefabcdefabcdefabcdefabcdefabcd)
+
+fetchcontent_makeavailable( gersemi )