+++ tests/formatter/issue_0021_builtin_commands_with_non_lower_case_canonical_names.out.cmake
@@ -1,0 +1,7 @@
+FetchContent_Declare(
+    gersemi
+    GIT_REPOSITORY https://github.com/BlankSpruce/gersemi.git
+    GIT_TAG abcdefabcdefabcdefabcdefabcdefabcdefabcd
+)
+
+FetchContent_MakeAvailable(gersemi)