+++ tests/formatter/issue_0021_custom_commands_with_non_lower_case_canonical_names.in.cmake
@@ -1,0 +1,46 @@
+### {definitions: [tests/formatter/issue_0021_custom_commands_with_non_lower_case_canonical_names.in.cmake]}
+function(CillianMurphy)
+    set(opts)
+    set(one_args)
+    set(multi_args)
+
+    cmake_parse_arguments(_ "${opts}" "${one_args}" "${multi_args}" ${ARGN})
+endfunction()
+
+function(bradleyCooper)
+    set(opts)
+    set(one_args)
+    set(multi_args)
+
+    cmake_parse_arguments(_ "${opts}" "${one_args}" "${multi_args}" ${ARGN})
+endfunction()
+
+function(COLMAN_DOMINGO)
+    set(opts)
+    set(one_args)
+    set(multi_args)
+
+    cmake_parse_arguments(_ "${opts}" "${one_args}" "${multi_args}" ${ARGN})
+endfunction()
+
+function(PaUlGiAmAtTi)
+    set(opts)
+    set(one_args)
+    set(multi_args)
+
+    cmake_parse_arguments(_ "${opts}" "${one_args}" "${multi_args}" ${ARGN})
+endfunction()
+
+function(jEfFrEy_wRiGhT)
+    set(opts)
+    set(one_args)
+    set(multi_args)
+
+    cmake_parse_arguments(_ "${opts}" "${one_args}" "${multi_args}" ${ARGN})
+endfunction()
+
+cillianmurphy()
+bradleycooper()
+colman_domingo()
+paulgiamatti()
+jeffrey_wright()