+++ tests/formatter/issue_0021_custom_commands_with_non_lower_case_canonical_names.out.cmake
@@ -1,0 +1,45 @@
+function(CillianMurphy)
+    set(opts)
+    set(one_args)
+    set(multi_args)
+
+    cmake_parse_arguments(_ "${opts}" "${one_args}" "${multi_args}" ${ARGN})
+endfunction()
+
+function(bradleyCooper)
+    set(opts)
+    set(one_args)
+    set(multi_args)
+
+    cmake_parse_arguments(_ "${opts}" "${one_args}" "${multi_args}" ${ARGN})
+endfunction()
+
+function(COLMAN_DOMINGO)
+    set(opts)
+    set(one_args)
+    set(multi_args)
+
+    cmake_parse_arguments(_ "${opts}" "${one_args}" "${multi_args}" ${ARGN})
+endfunction()
+
+function(PaUlGiAmAtTi)
+    set(opts)
+    set(one_args)
+    set(multi_args)
+
+    cmake_parse_arguments(_ "${opts}" "${one_args}" "${multi_args}" ${ARGN})
+endfunction()
+
+function(jEfFrEy_wRiGhT)
+    set(opts)
+    set(one_args)
+    set(multi_args)
+
+    cmake_parse_arguments(_ "${opts}" "${one_args}" "${multi_args}" ${ARGN})
+endfunction()
+
+CillianMurphy()
+bradleyCooper()
+COLMAN_DOMINGO()
+PaUlGiAmAtTi()
+jEfFrEy_wRiGhT()