+++ tests/formatter/set_package_properties.out.cmake
@@ -1,0 +1,11 @@
+set_package_properties(foo PROPERTIES URL foo)
+
+set_package_properties(
+    foo
+    PROPERTIES
+        URL foo______________________________
+        DESCRIPTION
+            bar____________________________________________________________
+        TYPE RECOMMENDED
+        PURPOSE baz______________________________
+)